+++ dev/e2e_app/lib/keys.dart
@@ -10,6 +10,12 @@ class Keys {
   static const bottomText = Key('bottomText');
   static const backButton = Key('backButton');
 
+  // at() finder screen
+  static const atFinderFirstItemTapped = Key('atFinderFirstItemTapped');
+  static const atFinderItem = Key('atFinderItem');
+  static const atFinderScreenButton = Key('atFinderScreenButton');
+  static const atFinderSecondItemTapped = Key('atFinderSecondItemTapped');
+
   // autofocus text field flow
   static const usernameTextField = Key('usernameTextField');
   static const usernameNextButton = Key('usernameNextButton');
@@ -30,6 +36,8 @@ class Keys {
   static const icon1 = Key('icon1');
   static const icon2 = Key('icon2');
   static const cameraFeaturesButton = Key('cameraFeaturesButton');
+  static const textfieldsScreenButton = Key('textfieldsScreenButton');
+  static const openExternalAppScreenButton = Key('openExternalAppScreenButton');
 
   // text fields screen
   static const textField1 = Key('textField1');
@@ -43,6 +51,7 @@ class Keys {
   static const microphonePermissionTile = Key('microphone');
   static const locationPermissionTile = Key('location');
   static const galleryPermissionTile = Key('gallery');
+  static const batteryPermissionTile = Key('battery');
   static const permissionsScreen = Key('permissionsScreen');
   static const requestCameraPermissionButton = Key(
     'requestCameraPermissionButton',
@@ -55,6 +64,9 @@ class Keys {
   );
   static const requestGalleryPermissionButton = Key(
     'requestGalleryPermissionButton',
+  );
+  static const requestBatteryPermissionButton = Key(
+    'requestBatteryPermissionButton',
   );
 
   // location screen
@@ -77,4 +89,27 @@ class Keys {
   static const pickMultiplePhotosButton = Key('pickMultiplePhotosButton');
   static const smallImagePreview = Key('smallImagePreview');
   static const selectedPhotosCount = Key('selectedPhotosCount');
+
+  // mobile keyboard enter test
+  static const textFields = [
+    Key('textField0'),
+    Key('textField1'),
+    Key('textField2'),
+    Key('textField3'),
+    Key('textField4'),
+    Key('textField5'),
+    Key('textField6'),
+    Key('textField7'),
+    Key('textField8'),
+    Key('textField9'),
+    Key('textField10'),
+  ];
+
+  // external app screen
+  static const openIosSettingsButton = Key('openIosSettingsButton');
+  static const externalAppStatusText = Key('externalAppStatusText');
+}
+
+class AtFinderItemKey extends ValueKey<int> {
+  const AtFinderItemKey(super.value);
 }